github-prs/requirements.txt
fastapi
httpx
PyJWT
uvicorn
python-dotenv
